Move EditorControlBar and EditorControlButton to the correct namespace
