+++ Quaver.Shared/Screens/Editor/UI/EditorControlBar.cs
@@ -1,16 +1,14 @@
-﻿using System;
-using Microsoft.Xna.Framework;
+﻿using Microsoft.Xna.Framework;
 using Quaver.Shared.Assets;
 using Quaver.Shared.Audio;
-using Quaver.Shared.Graphics;
 using Quaver.Shared.Graphics.Notifications;
-using Quaver.Shared.Helpers;
+using Quaver.Shared.Screens.Editor.UI.Rulesets;
 using Quaver.Shared.Screens.Menu.UI.Jukebox;
 using Wobble.Graphics;
 using Wobble.Graphics.Sprites;
 using Wobble.Window;
 
-namespace Quaver.Shared.Screens.Editor.UI.Rulesets
+namespace Quaver.Shared.Screens.Editor.UI
 {
     public class EditorControlBar : Sprite
     {
